Добавляет Модуляцию
